FEAN-256 [FEAT] 프론트 연결 작업 및 쿠폰 발급 로직 수정, CORS 해결을 위한 설정 추가
