+++ src/main/java/com/lotte/danuri/promotion/filter/CustomHandlerFilterFunction.java
@@ -1,11 +1,16 @@
 package com.lotte.danuri.promotion.filter;
 
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.reactive.function.server.HandlerFilterFunction;
 import org.springframework.web.reactive.function.server.HandlerFunction;
 import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
 import reactor.core.publisher.Mono;
 
+@CrossOrigin("*")
+@Slf4j
 public class CustomHandlerFilterFunction implements HandlerFilterFunction<ServerResponse, ServerResponse> {
 
     @Override
