+++ src/main/java/com/lotte/danuri/promotion/handler/PromotionHandler.java
@@ -39,19 +39,11 @@ public class PromotionHandler {
         String memberId = request.headers().header("memberId").get(0);
         //String promotionId = request.headers().header("promotionId").get(0);
 
-        //String memberId = "1";
         // -1 : 프로모션 종료, null : 대기열 존재 안함, 작업열 존재 -> 쿠폰 받은 거 성공
         Long rank = redisService.getOrderNumber(Promotion.PROMOTION.waitKey, memberId);
         if(rank == null) { // 대기열 존재 안할 때 -> 1. 작업열로 이동, 2. 대기열에 진입 조차 못함
-            Long workRank = redisService.getWorkNumber(Promotion.PROMOTION.workKey, memberId);
-
-            if(workRank != null) {
-                String msg = "프로모션 응모에 성공하셨습니다!";
-                return ServerResponse.ok().contentType(MediaType.APPLICATION_JSON).bodyValue(msg);
-            }
-
-            rank = -100L;
-
+            String msg = "success";
+            return ServerResponse.ok().contentType(MediaType.APPLICATION_JSON).bodyValue(msg);
         }
 
         return ok().contentType(MediaType.APPLICATION_JSON).bodyValue(rank);
