+++ src/main/java/com/lotte/danuri/promotion/scheduler/QueueScheduler.java
@@ -23,15 +23,15 @@ public class QueueScheduler {
             log.info("======= 프로모션이 종료되었습니다. =======");
             log.info("작업열 사이즈 = {}",redisService.getSizeOfWork(Promotion.PROMOTION));
 
-            return;
-        }
-        if(redisService.getSize(Promotion.PROMOTION) >= Promotion.PROMOTION.limit) {
-            redisService.move(Promotion.PROMOTION);
-
-            redisService.publish(Promotion.PROMOTION);
-
             redisService.delete(Promotion.PROMOTION);
             redisService.setPromotionCount(Promotion.PROMOTION.limit);
+
+            return;
+        }
+
+        if(redisService.getSize(Promotion.PROMOTION) > 0) {
+            redisService.move(Promotion.PROMOTION);
+            redisService.publish(Promotion.PROMOTION);
         }
 
     }
